feat: Add Commitlint Step for Commit Message Validation
- Introduced a new Commitlint step in the pipeline to enforce commit message standards.
- Updated Steps.pkl to include the new Commitlint step.
