+++ .pre-commit-config.pkl
@@ -72,7 +72,8 @@ local function normalizeList(value: Any?): List =
 // - hk batch=false (default): run command once with all files
 //   -> pre-commit pass_filenames=true if {{files}} in command, false otherwise
 // - hk {{files}} template -> remove it, pre-commit handles file passing
-local function stepToHook(id: String, step: Any): Dynamic? =
+// - hk {{commit_msg_file}} template -> map to PRE_COMMIT_COMMIT_MSG env fallback
+local function stepToHook(hookName: String, id: String, step: Any): Dynamic? =
   let (stepDyn = (step as Typed).toDynamic())
   let (profiles = normalizeList(stepDyn.profiles))
     if (
@@ -95,14 +96,15 @@ local function stepToHook(id: String, step: Any): Dynamic? =
       let (prefix = if (stepDyn.prefix != null) stepDyn.prefix else "")
       let (hasFilesPlaceholder = checkCmd.contains("{{files}}") || checkCmd.contains("{{ files }}"))
       let (rawEntry = if (prefix != "") prefix + " " + checkCmd else checkCmd)
+      let (commitMsgPlaceholder = "\"${PRE_COMMIT_COMMIT_MSG:-.git/COMMIT_EDITMSG}\"")
       // Remove {{files}} placeholders - pre-commit handles file passing
-      // Also remove shell redirections (>, >>, 2>&1, etc.) since pre-commit
-      // appends files at the end
       let (
         entry =
           rawEntry
             .replaceAll("{{files}}", "")
             .replaceAll("{{ files }}", "")
+            .replaceAll("{{commit_msg_file}}", commitMsgPlaceholder)
+            .replaceAll("{{ commit_msg_file }}", commitMsgPlaceholder)
             .replaceAll(">/dev/null", "")
             .replaceAll("2>/dev/null", "")
             .replaceAll("2>&1", "")
@@ -112,7 +114,13 @@ local function stepToHook(id: String, step: Any): Dynamic? =
       // Determine pass_filenames based on whether command expects files
       // Key rule: pass filenames when batch mode is enabled or {{files}} placeholder was used
       let (isBatch = stepDyn.batch == true)
-      let (passFilenames = isBatch || hasFilesPlaceholder)
+      let (
+        passFilenames =
+          if (hookName == "commit-msg")
+            false
+          else
+            isBatch || hasFilesPlaceholder
+      )
       let (stagePatterns = normalizeList(stepDyn.stage))
       let (globPatterns = normalizeList(stepDyn.glob))
       let (
@@ -137,31 +145,38 @@ local function stepToHook(id: String, step: Any): Dynamic? =
           when (!excludePatterns.isEmpty) {
             ["exclude"] = excludeListToRegex(excludePatterns)
           }
+          when (hookName == "commit-msg") {
+            ["stages"] = List("commit-msg")
+          }
         }
     else
       null
 
 // Process steps from a group or single step
-local function processItem(name: String, item: Any): List =
+local function processItem(hookName: String, name: String, item: Any): List =
   let (itemDyn = (item as Typed).toDynamic())
     if (itemDyn._type == "step")
-      let (hook = stepToHook(name, item))
+      let (hook = stepToHook(hookName, name, item))
         if (hook != null) List(hook) else List()
     else if (itemDyn._type == "group")
       itemDyn.steps.keys
         .toList()
         .flatMap((stepName) ->
           let (step = itemDyn.steps[stepName])
-          let (hook = stepToHook(stepName, step))
+          let (hook = stepToHook(hookName, stepName, step))
             if (hook != null) List(hook) else List()
         )
     else
       List()
 
-// Process all steps from hooks
-local preCommitHook = hk.hooks["pre-commit"]
+// Process all supported hooks
+local hookNames = List("pre-commit", "commit-msg")
+
 local allHooks: List =
-  preCommitHook.steps.keys.toList().flatMap((name) -> processItem(name, preCommitHook.steps[name]))
+  hookNames.flatMap((hookName) ->
+    let (hook = hk.hooks[hookName])
+      hook.steps.keys.toList().flatMap((name) -> processItem(hookName, name, hook.steps[name]))
+  )
 
 // Output pre-commit config
 output {
