+++ package.json
@@ -16,6 +16,11 @@
   },
   "devDependencies": {
     "@biomejs/biome": "2.3.7",
+    "@commitlint/cli": "^20.1.0",
+    "@commitlint/config-conventional": "^20.0.0",
+    "@commitlint/format": "^20.0.0",
+    "@commitlint/types": "^20.0.0",
+    "ap-style-title-case": "^2.0.0",
     "prettier": "3.6.2"
   }
 }
