+++ src/private/hk/steps/Commitlint.pkl
@@ -1,0 +1,6 @@
+import "@hk/Config.pkl"
+
+commitlint = new Config.Step {
+  prefix = "pnpm exec --"
+  check = "commitlint --strict --edit"
+}
